distribution: Add in-scope and out-of-scope wording
Docker's use of Bearer requires information beyond what's covered in
RFC 6749 and 6750 [1].  Folks writing a client that will interact with
a Docker registry that uses that auth approach will need a "Docker
registry's 'Bearer' additions" spec to follow, but Derek believes the
situation is salvageable with external work [2].

Also pin the docker/registry links to a specific version so the links
will survive future docker/registry changes (including removing the
docs after the OCI picks them up).  As long as the TOB-selected
version isn't far behind (how far will the spec move during a week of
voting?), it should be easy for the new maintainets to catch up on any
subsequent drift.

The signing scope is from Stephen in [3].  The discovery scope is from
Derek [4].  The content-management scope is from Stephen [5] and Liang
[6].

This commit also add's Mike's interoperability statement [7], which
mentions one reason for the OCI inclusion is the
implementation-agnostic location where implementers can collaborating
on a common specification.  The project scoping is open to drift with
the limits imposed by the TOB's scope table.  If the TOB thinks
subsequent drift is excessive, they are free to make further
scope-table adjustments in follow-up proposals.

I've also added Mike's library/busybox scoping example [8].

[1]: https://github.com/xiekeyang/oci-discovery/pull/64#issue-291807003
[2]: https://github.com/opencontainers/tob/pull/37#issuecomment-360923589
[3]: https://github.com/opencontainers/tob/pull/35#discussion_r164012767
[4]: https://github.com/opencontainers/tob/issues/34#issuecomment-350529321
[5]: https://github.com/opencontainers/tob/pull/35#discussion_r164012767
[6]: https://github.com/opencontainers/tob/pull/37#issuecomment-360968873
[7]: https://github.com/opencontainers/tob/pull/35#discussion_r163644412
[8]: https://github.com/opencontainers/tob/pull/37#discussion_r164564620

Signed-off-by: W. Trevor King <wking@tremily.us>