distribution: Copy-edits for the scope table
There's precendent in the scope table for "Not currently being worked"
(the "Creating Reference spec for optional DNS based naming &
distribution" entry [1]), but "Work not yet started" (from the
"Archival Format" entry [1]) is a more complete fragment.

The "...provide a standardized way to use DNS..." line is part of the
"Creating a reference spec for optional DNS based naming and
discovery" entry, so I'm dropping it from the "Specifying
authentication and authorization schemes" entry.

I've dropped one of the "parallel" instances for from the
parallel-naming/discovery sentence.  No need to say that twice in once
sentence.

[1]: https://www.opencontainers.org/about/oci-scope-table

Signed-off-by: W. Trevor King <wking@tremily.us>