README: Use relative links to the proposals
There's no need to use absolute links for these.  And using relative
links sets a good example for future proposals, because the links will
work in the proposal branch's README before the referenced proposal
lands in the TOB master.

There's a similar absolute link for the code-of-conduct in the
distribution proposal.  The intention may have been to show the
absolute URL which would be included in new project's repository.  I
don't think that matters, because the project template has included a
similar link since opencontainers/project-template@ee72bc89
(CONTRIBUTING: Code of conduct, meetings, mailing list, and IRC,
2016-09-08, #45), so I've left that alone here.

Signed-off-by: W. Trevor King <wking@tremily.us>