Update charter to remove scope table
The OCI scope table as an artifact when the OCI was bootstrapped as an organization. The scope of the OCI has expanded over time to include image and distribution specifications. At the end of the day, the TOB controls the overall technical scope of the organization which is fairly well defined in the charter already as "The Open Container Initiative provides an open source technical community within which industry participants may easily contribute to building a vendor-neutral, portable and open specifications and runtime that deliver on the promise of containers as a source of application portability backed by a certification program."

Signed-off-by: Chris Aniszczyk <caniszczyk@linuxfoundation.org>