CHARTER: word-wrap at 80 characters
This doesn't contain any text changes, it only changes the layout so
that you can effectively read the charter in a terminal. I know it is
"bad practice" in the OCI to have hard line-wrapping (the argument being
that it is harder to review changes), but I've never found that to be an
issue.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>