CHARTER: s12(c): formalise charter versions and changelog
We have version numbers (and a changelog) in the Charter, so we should
probably actually enshrine that in the Charter. In addition, clean up
the changelog format so that it's easier to understand.

This change does introduce new requirements for the editors of the
Charter, but it simply mirrors the current convention applied to Charter
amendments.

In addition, define the concept of a "draft Charter version" so that we
can make iterative changes to the charter through separate PRs without
having to enact each minor change separately.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>