CHARTER: s6(n): clarify and unify TOB voting rules
Section 6 (n) conflicted with all of the other mentions of TOB votes and
TOB decisions, and we have always required a qualified super-majority
for all TOB votes -- so it is best to properly codify that requirement
in the Charter.

In addition, Section 6 (n) appears to have had a copy-paste error which
made the paragraph nonsensical (the Trademark Board has nothing to do
with TOB votes). This has also been corrected.

In addition, it is unnecessary to have a special carve-out for changing
voting systems in Section 6 (h) because the TOB can simply amend the
Charter to change the electoral system for TOB elections (which requires
the same qualified super-majority -- and would likely be needed purely
to allow TOB candidates to understand how the voting system operates).

This change does (on paper) change the procedure for TOB votes, but the
intention is to simply mirror the current convention of the TOB when it
comes to voting on motions. It also removes an apparent mistake in the
original Charter, as I do not believe it was ever intended that the
Trademark Board be involved in TOB decisions.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>