CHARTER: s2: refine OCI Projects definition and scope
We already have more than one type of project, but Section 2 (as
written) doesn't appear to have a clear way of describing the different
types of projects that exist. Given the "getting started" document's
selection of "project buckets", we need to reflect those views in the
Charter.

This change does rearrange some parts of this Section of the Charter
(and introduces several new concepts), but it is intended to primarily
mirror the existing conventions of the OCI. In fact, it legitimises most
of the projects which have been included after the initial Charter came
into effect.

This change *does not* add any additional restrictions to existing
projects or potential OCI projects, as that will be handled in a
separate patch. An additional restriction is added to OCI Conformance
Suites, but that was a restriction which had already been followed in
practice by the respective OCI Conformance Suite projects, so it simply
is mirroring existing convention.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>