CHARTER: s2(h)..(j): restrict OCI Reference Implementations scope
A common problem we've had with runc is that we have ended up
implementing a lot of features that are technically within the scope of
the OCI Runtime Specification but are not defined by that specification.
As a result, users end up depending on runc's particular behaviour
rather than the behaviour outlined in the specification.

This hopefully establishes a middle-ground between allowing for
experimental development within an OCI Reference Implementation, and the
anarchy that is allowing for features to become depended on within an
OCI Reference Implementation without the prerequisite OCI Specification
work.

Unfortunately, this problem already exists within runc so we cannot
mandate it to no longer be the case (sadly that's not how technology
works). But the next best thing we can do is mandate it for any future
OCI Reference Implementations, and provide a one-time carve-out for the
runc project (to avoid this Charter update from bringing into question
the validity of the runc project within the OCI).

This change does introduce new restrictions for OCI Reference
Implementations which were not an existing convention for runc, though
in my view this would've been the convention for future OCI Reference
Implementations without this Charter change.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>