CHARTER: s5: rework TDC description and scope
The TDC has been a fairly confusing concept from the outset, and the
statement of the scope of the TDC was quite difficult to follow (not to
mention that it contained some text which was only relevant to the
runtime-spec). The text has been refreshed and now has a more general
description of the role of the TDC now that the OCI has a pretty large
variety of projects under its belt, and explicitly defines the
Maintainership Coalition (the set of all maintainers across all
projects). Previously, a literal reading of the Charter would have you
believe that only runtime-spec maintainers had certain rights under the
Charter.

This change does redefine some aspects of the TDC, and technically
changes the roles of different TDC elements, but broadly this is in
keeping with existing conventions within the TDC of OCI Projects and
doesn't majorly change the role of the TDC within the OCI.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>