CHARTER: s6(g)(5): exclude incapable TOB members-elect
If a TOB member-elect becomes incapable of being a TOB member after the
election, exclude them from the members-elect set and hold a by-election
after the TOB seats expire.

This avoids the problem of expelling 3 or more TOB members after an
election due to a merger, and simply pretends as though the member had
become inelligible before the election began. This is simpler and more
fair to both voters and the members-elect than re-running the entire
election or simulating a re-run of the election using the Condorcet
results.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>